Extrator HTML conectado com a interface
