+++ src/parsing/html/parsing_html_div.py
@@ -6,70 +6,6 @@ import xml
 import re
 import errno
 
-
-def clean_html(html_file, isString):
-    '''
-    Receives the html file and removes unecessery parts, as header, footer, etc.
-    '''
-
-    # List of elements that are going to be removed from the html
-    remove_list = ["head", "header", "footer", "polygon", "path", "script",
-                   "symbol", "meta", "link", "title", "style", "nav", "table",
-                   "form"]
-    remove_class = ["sidebar-inner", "breadcrumb", "share", "navegacao",
-                    "skiptranslate", "goog-te-spinner-pos", "social-list",
-                    "social-icon", "copyright", "id_assist_frame",
-                    "fbc-badge-tooltip"]
-    remove_id = ["boxes", "mySidenav", "chat-panel"]
-
-    # Check if the html_file is a string with the page or a path to the file
-    if isString:
-        f = html_file
-        soup = BeautifulSoup(f, 'html.parser')
-    else:
-        soup = ""
-        f = open(html_file, encoding="ISO-8859-1")
-        soup = BeautifulSoup(f, 'html.parser')
-        f.close()
-
-
-    # Remove any tag present in remove_list
-    for tag in soup.find_all():
-        if tag.name.lower() in remove_list:
-            tag.extract()
-    # Remove any div with the class in remove_class
-    for div in soup.find_all("div", {'class': remove_class}):
-        div.extract()
-    # Remove any div with the id in remove_id
-    for div in soup.find_all("div", {'id': remove_id}):
-        div.extract()
-
-    html_file = str(soup)
-
-
-
-    return html_file
-
-
-def fix_links(html_file):
-    '''
-    Receives the html file and return the same file with the value of the links
-    as the text of the links
-    '''
-    f = html_file
-    soup = BeautifulSoup(f, 'html.parser')
-
-    for a in soup.find_all('a', href=True):
-        if "http" in a['href']:
-            a.string = a['href']
-
-    html_file = str(soup)
-
-
-
-    return html_file
-
-
 def extrac_div(html_file):
     '''
     Receives a html file and creates a list of elements with the content of the
@@ -91,12 +27,12 @@ def write_csv(csv_list, csv_output_name):
     '''
     Receives a list of content and the name of the csv file, saves the csv file
     '''
-    with open(csv_output_name, "w", newline="") as f:
+    with open(csv_output_name, "a", newline="") as f:
         writer = csv.writer(f)
         writer.writerow(csv_list)
 
 
-def html_to_csv(html_file_path, is_string=False, output_file='output.csv'):
+def div_to_csv(html_file_path, is_string=False, output_file='output.csv'):
     '''
     Receives an html file path, converts the html to csv and saves the file on
     disk.
@@ -108,16 +44,10 @@ string or as the path to the file)
     :param output_file : str, optional (Name and path of the output csv file,
     default is output.csv)
     '''
-
     # Check if html file exists
     if (os.path.isfile(html_file_path)) or is_string:
-
-        # Clean the html file
-        html_file = clean_html(html_file_path, is_string)
-        # Fix the links in the file
-        html_file = fix_links(html_file)
         # Extract the content
-        csv_list_all = extrac_div(html_file)
+        csv_list_all = extrac_div(html_file_path)
         # Saves the content in a csv file
         write_csv(csv_list_all, output_file)
